Exeption handling

+++ dscommerce/src/main/java/com/lgomeszx/dscommerce/services/ProductService.java
@@ -3,12 +3,16 @@ package com.lgomeszx.dscommerce.services;
 import com.lgomeszx.dscommerce.dto.ProductDTO;
 import com.lgomeszx.dscommerce.entities.Product;
 import com.lgomeszx.dscommerce.repositories.ProductRepository;
+import com.lgomeszx.dscommerce.services.exeptions.DatabaseException;
+import com.lgomeszx.dscommerce.services.exeptions.ResorceNotFoundExeption;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.DeleteMapping;
 
 import java.util.List;
 
@@ -20,7 +24,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResorceNotFoundExeption("Recurso nao encontrado."));
         ProductDTO dto = new ProductDTO(product);
         return dto;
     }
@@ -34,30 +39,41 @@ public class ProductService {
     @Transactional
     public ProductDTO insert(ProductDTO dto) {
         Product entity = new Product();
-        entity.setName(dto.getName());
-        entity.setDescription(dto.getDescription());
-        entity.setPrice(dto.getPrice());
-        entity.setImgUrl(dto.getImgUrl());
-
+        setProduct(dto, entity);
         entity = repository.save(entity);
         return new ProductDTO(entity);
     }
 
     @Transactional
     public ProductDTO update(Long id,ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
+        try {
+            Product entity = repository.getReferenceById(id);
+            setProduct(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResorceNotFoundExeption("Recurso nao encontrado.");
+        }
+    }
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResorceNotFoundExeption("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+    }
+
+    private void setProduct(ProductDTO dto, Product entity)  {
         entity.setName(dto.getName());
         entity.setDescription(dto.getDescription());
         entity.setPrice(dto.getPrice());
         entity.setImgUrl(dto.getImgUrl());
-
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
-    }
-
-    @Transactional
-    public void delete(Long id) {
-        repository.deleteById(id);
     }
 
 }
